feat: 조인 테이블 어노테이션 추가, event 패키지 생성
