+++ src/main/java/com/happiday/Happi_Day/domain/entity/event/EventComment.java
@@ -1,0 +1,36 @@
+package com.happiday.Happi_Day.domain.entity.event;
+
+import com.happiday.Happi_Day.domain.entity.BaseEntity;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@SuperBuilder(toBuilder = true)
+@Table(name ="event_comment")
+public class EventComment extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    // 유저와 관계 설정
+//    @ManyToOne(fetch = FetchType.LAZY)
+//    @JoinColumn(name = "user_id")
+//    private UserEntity user;
+
+    // 이벤트 게시글과 관계 설정
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "event_id")
+    private Event event;
+
+    @NotNull
+    @Column(columnDefinition = "TEXT")
+    private String content;
+}
